+++ src/oaklib/implementations/ontobee/ontobee_implementation.py
@@ -2,7 +2,7 @@ from abc import ABC
 from dataclasses import dataclass
 from typing import Optional
 
-from oaklib.implementations.sparql.sparql_implementation import SparqlImplementation
+from oaklib.implementations.sparql.abstract_sparql_implementation import AbstractSparqlImplementation
 from oaklib.interfaces.mapping_provider_interface import MappingProviderInterface
 from oaklib.interfaces.obograph_interface import OboGraphInterface
 from oaklib.interfaces.search_interface import SearchInterface
@@ -10,7 +10,7 @@ from oaklib.interfaces.search_interface import SearchInterface
 ONTOBEE_MERGED_GRAPH_PREFIX = 'http://purl.obolibrary.org/obo/merged/'
 
 @dataclass
-class OntobeeImplementation(SparqlImplementation, SearchInterface, MappingProviderInterface, OboGraphInterface):
+class OntobeeImplementation(AbstractSparqlImplementation, SearchInterface, MappingProviderInterface, OboGraphInterface):
     """
     Wraps the Ontobee sparql endpoint
