+++ tests/__init__.py
@@ -22,6 +22,7 @@ CELLULAR_ANATOMICAL_ENTITY = 'GO:0110165'
 CELL = 'CL:0000000'
 SHAPE = 'PATO:0000052'
 PHOTORECEPTOR_OUTER_SEGMENT = 'GO:0001750'
+CATALYTIC_ACTIVITY = 'GO:0003824'
 
 CHEBI_NUCLEUS = 'CHEBI:33252'
 NUCLEUS = 'GO:0005634'
